Updated Object Types
- Updated types to match PR #1073
- Part of the motivaiton for this is that `/dairy/services.js` also
relies on unifiedDataLoader; if this is written using those
types, then that rewrite can build off of this rewrite without
any additional merge conflicts down the road.
