--- www/js/types/fileShareTypes.ts
@@ -1,12 +1,0 @@
-export interface FsWindow extends Window {
-  requestFileSystem: (
-    type: number,
-    size: number,
-    successCallback: (fs: any) => void,
-    errorCallback?: (error: any) => void
-  ) => void;
-  LocalFileSystem: {
-    TEMPORARY: number;
-    PERSISTENT: number;
-  };
-};
